# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "校园失物招领智能匹配平台 API",
    "description": "《软件系统分析与设计综合实践》课程项目接口契约。统一响应结构：{ code: 0成功/非0失败, message: 提示信息, data: 业务数据 }。除注册登录外，所有接口需在 Header 携带 Authorization: Bearer {token}。",
    "version": "v1.0 (第3周契约初稿)"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "后端本地开发" }
  ],
...
  "paths": {
...
    "/api/items": {
      "post": {
        "tags": ["信息"],
        "summary": "发布失物/招领信息（招领需带隐藏特征）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ItemCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发布成功；系统异步触发智能匹配，命中后通过站内信推送",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" }, "example": { "code": 0, "message": "发布成功，匹配进行中，命中后将通知您", "data": { "id": 20001 } } } }
          }
        }
      },
      "get": {
        "tags": ["信息"],
        "summary": "信息流检索（多条件 + 分页）",
        "parameters": [
          { "name": "type", "in": "query", "required": true, "schema": { "type": "string", "enum": ["LOST", "FOUND"] } },
          { "name": "categoryId", "in": "query", "schema": { "type": "integer" } },
          { "name": "locationId", "in": "query", "schema": { "type": "integer" } },
          { "name": "keyword", "in": "query", "schema": { "type": "string" }, "description": "按标题/描述模糊搜索" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" }, "example": { "code": 0, "message": "success", "data": { "total": 57, "page": 1, "size": 10, "list": [ { "id": 20001, "type": "FOUND", "title": "黑色皮质卡包", "categoryId": 2, "categoryName": "钱包", "locationName": "清真食堂门口", "eventTime": "2026-09-10 12:30:00", "coverImage": "/files/img/f001.jpg", "status": "OPEN", "createdAt": "2026-09-10 13:02:11" } ] } } } }
          }
        }
      }
    },
...
    "/api/claims": {
      "post": {
        "tags": ["认领"],
        "summary": "提交认领申请（需回答隐藏特征）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["foundItemId", "answers"],
                "properties": {
                  "foundItemId": { "type": "integer", "example": 20001 },
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "featureKey": { "type": "string", "example": "卡内姓名" },
                        "answer": { "type": "string", "example": "张三" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "提交成功，进入待核验", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" }, "example": { "code": 0, "message": "申请已提交，等待拾获者核验", "data": { "claimId": 40001 } } } } } }
      }
    },
...
    "/api/admin/items": {
      "get": {
        "tags": ["后台"],
        "summary": "待审核信息列表（管理员/审核员）",
        "parameters": [ { "name": "status", "in": "query", "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
    "/api/admin/users/{id}/ban": {
      "post": {
        "tags": ["后台"],
        "summary": "封禁/解封用户",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "ban": { "type": "boolean" }, "reason": { "type": "string" } }, "required": ["ban"] } } } },
        "responses": { "200": { "description": "成功", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
    "/api/admin/claims": {
      "get": {
        "tags": ["后台"],
        "summary": "争议认领单队列（默认 DISPUTED）",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "default": "DISPUTED" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "成功；含双方信息、特征作答明细（matched 标记）、争议原因",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" }, "example": { "code": 0, "message": "success", "data": { "total": 1, "page": 1, "size": 10, "list": [ { "id": 40001, "foundItem": { "id": 20001, "title": "黑色皮质卡包" }, "claimant": { "id": 10002, "username": "zhangsan", "nickname": "张三", "creditScore": 100 }, "publisher": { "id": 10003, "username": "lisi", "nickname": "李四", "creditScore": 105 }, "featureAnswers": [ { "featureKey": "卡内姓名", "answer": "张三", "matched": true } ], "disputeReason": "回答正确但拾获者存疑", "status": "DISPUTED", "createdAt": "2026-09-12 10:00:00" } ] } } } }
          }
        }
      }
    },
    "/api/admin/claims/{id}/arbitrate": {
      "post": {
        "tags": ["后台"],
        "summary": "争议仲裁（支持认领人/驳回）",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["approve", "reason"],
                "properties": {
                  "approve": { "type": "boolean", "description": "true=支持认领人（APPROVED 并生成核销码）；false=驳回（REJECTED，认领人信用分 -20）" },
                  "reason": { "type": "string", "description": "仲裁理由，必填" }
                },
                "example": { "approve": false, "reason": "特征作答不足以证明归属" }
              }
            }
          }
        },
        "responses": { "200": { "description": "仲裁完成：状态切换、信用分/核销码联动、通知双方、写 audit_logs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer", "description": "0=成功，非0=业务错误码" },
          "message": { "type": "string" },
          "data": { "type": "object", "nullable": true }
        }
      },
...
      "ItemCreateRequest": {
        "type": "object",
        "required": ["type", "title", "categoryId", "eventTime", "description"],
        "properties": {
          "type": { "type": "string", "enum": ["LOST", "FOUND"] },
          "title": { "type": "string", "maxLength": 100, "example": "黑色皮质卡包" },
          "categoryId": { "type": "integer", "example": 2 },
          "locationId": { "type": "integer", "nullable": true, "description": "记不清可不传", "example": 14 },
          "eventTime": { "type": "string", "format": "date-time", "example": "2026-09-10 12:30:00" },
          "description": { "type": "string", "example": "在清真食堂门口捡到一个黑色皮质卡包" },
          "images": { "type": "array", "items": { "type": "string" }, "description": "先调 /api/files 上传得到的路径", "example": ["/files/img/abc.jpg"] },
          "hiddenFeatures": {
            "type": "array",
            "description": "仅 FOUND 类型必填：防冒领的隐藏特征，不公开展示",
            "items": {
              "type": "object",
              "properties": {
                "featureKey": { "type": "string", "example": "卡内姓名" },
                "answer": { "type": "string", "example": "张三" }
              }
            }
          }
        }
      }
    }
  }
}